fix(#273 T4): review round 3 mechanical items — bypass-injection coverage, file split, tombstone comment, Logfire warning
Two mechanical fixes plus three follow-ups (both seats approved conceptually,
these were the remaining machinery per the coordinator):

- Backstop coverage (patch coverage 86% -> restored): `_evict_by_budget`,
  `_drop_oldest`, and `_evict_by_count`'s unconditional trailing trim were
  arithmetically unreachable via the public write path (96B x 8 x 2 <= 4147B
  << 8192B budget; the write path always supersedes the whole prior live
  set before appending, so "too few superseded to reach the count cap" can't
  arise there either). Two new tests inject oversized/under-superseded
  records directly via `FactLedger.model_validate` / direct construction —
  bypassing the normal write-path truncation and supersession — which is
  exactly the scenario these two backstops exist for, and now exercises
  both lines.
- Split `test_fact_ledger.py` (was 254 lines) into `test_fact_ledger.py`
  (core append/supersede/bound/hygiene, 191 lines) and
  `test_fact_ledger_envelope.py` (consumption gate, envelope serialization,
  rollback parse, 154+ lines) per the repo's ≤200-line test-file rule.
- `_TURN_SUPERSEDED`'s comment now states plainly that it is an intentional,
  permanent tombstone (a whole-set turn replacement has no 1:1 old-record ->
  new-record correspondence, so there is no single "successor" id to write)
  rather than the previous, inaccurate "real ids replace it once eviction
  runs."
- `_parse_forward_compatible` now emits `logger.warning("fact_ledger_dropped",
  dropped_keys=...)` (Logfire-visible) whenever the rollback-compat path
  actually engages, so a rollback silently degrading fact-ledger content is
  observable rather than invisible. New test mocks `session_facade.logger`
  and asserts the call.
- New split-codepoint truncation test: `_truncate`'s byte-aware cut must
  never leave a mangled partial UTF-8 sequence (CJK characters are 3 bytes).

Mutation-verified (applied then reverted): no-op'ing the trailing trim in
`_evict_by_count` fails the new backstop test (10 records survive instead
of 8); no-op'ing `_evict_by_budget` fails the new budget test (17020 bytes
survive instead of dropping under 8192); removing the `logger.warning` call
fails the new Logfire-visibility test.

Gates: make lint / make typecheck clean; make test 1521 passed, 89.30%
coverage (floor 87%, up from 89.10%).
